Update on "[coor] Represent get_rank as a SymInt"
compile_on_one_rank assumes rank 0 can produce a graph that is valid across the group. Dynamo normally constant folds torch.distributed.get_rank(), so rank-dependent Python control flow can be baked into the graph before the artifact is reused on other ranks.

Instead, default get_rank() under compile_on_one_rank is traced as a runtime SymInt via a small device_mesh op. Rank arithmetic can stay in the graph, while Python control flow such as if get_rank() == 0 now fails through the normal data-dependent symbolic guard path.

cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx kadeng chauhang amjames jataylo azahed98

[ghstack-poisoned]
