+++ torch/_dynamo/trace_rules.py
@@ -173,10 +173,11 @@ manual_torch_name_rule_map: dict[
     "torch.jit.annotate": TorchInGraphFunctionVariable,
     "torch.distributed.is_available": TorchInGraphFunctionVariable,
     "torch.distributed.is_initialized": TorchInGraphFunctionVariable,
-    "torch.distributed.get_rank": TorchInGraphFunctionVariable,
+    "torch.distributed.get_rank": UserFunctionVariable,
     "torch.distributed.get_world_size": TorchInGraphFunctionVariable,
     "torch.distributed.tensor._api.DTensor#from_local": TorchInGraphFunctionVariable,
     "torch.distributed.device_mesh.DeviceMesh#__init__": SkipFunctionVariable,
+    "torch.distributed.distributed_c10d._get_rank": TorchInGraphFunctionVariable,
     "torch.distributed.distributed_c10d._get_group_size_by_name": TorchInGraphFunctionVariable,
     "torch.distributed.distributed_c10d._resolve_group_name_by_ranks_and_tag": TorchInGraphFunctionVariable,
     "torch.distributed.distributed_c10d._get_group_tag": TorchInGraphFunctionVariable,
