+++ torch/distributed/distributed_c10d.py
@@ -2547,7 +2547,18 @@ def _abort_process_group(group: ProcessGroup | None = None):
         _unregister_process_group(pg.group_name)
 
 
-def get_rank(group: ProcessGroup | None = None) -> int:
+def _get_rank(group: ProcessGroup | None = None) -> int:
+    if _rank_not_in_group(group):
+        return -1
+
+    default_pg = _get_default_group()
+    if group is None or group is GroupMember.WORLD:
+        return default_pg.rank()
+
+    return get_group_rank(group, default_pg.rank())
+
+
+def get_rank(group: ProcessGroup | None = None) -> int | torch.SymInt:
     """
     Return the rank of the current process in the provided ``group``, default otherwise.
 
@@ -2564,14 +2575,13 @@ def get_rank(group: ProcessGroup | None = None) -> int:
         -1, if not part of the group
 
     """
-    if _rank_not_in_group(group):
-        return -1
-
-    default_pg = _get_default_group()
     if group is None or group is GroupMember.WORLD:
-        return default_pg.rank()
+        if dist_config.compile_on_one_rank:
+            from torch.distributed._ops import _runtime_rank  # noqa: F401
 
-    return get_group_rank(group, default_pg.rank())
+            return torch.ops.device_mesh._runtime_get_rank.default()
+
+    return _get_rank(group)
 
 
 def get_world_size(group: ProcessGroup | None = None) -> int:
